safari-600.7-branch/LayoutTests/fast/dom/DOMURL/url-password-expected.txt

Test getting the password attribute of a URL object.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Default password is empty
PASS a.password is ''
Unspecified password should return empty string
PASS a.password is ''
Empty password should be empty
PASS a.password is ''
Regular passowrd retrieval
PASS a.password is 'abc123'
Password with empty username
PASS a.password is 'def456'
Setting password
PASS a.href is 'http://:xyzzy@example.com/'
PASS successfullyParsed is true

TEST COMPLETE
